replace float with double in low level objects and operations/new primitive: Plane

+++ rtmlib/include/computations/computation.h
@@ -9,7 +9,7 @@ namespace rtm
 {
 struct Computation
 {
-    float tvalue;
+    double tvalue;
     Surface* surface;
     Point point;
     Vector eyev;
