+++ rtmlib/include/geometry/intersection.h
@@ -10,14 +10,14 @@ namespace rtm
 class Intersection : public Intersections
 {
     private:
-        float m_t;
+        double m_t;
         Surface* m_object;
     public:
         Intersection();
 
-        Intersection(const float& t_val , const Surface* object_val);
+        Intersection(const double& t_val , const Surface* object_val);
 
-        float get_t() const override;
+        double get_t() const override;
         Surface* get_shape() const override; 
 
         // Useful for std::sort array of intersections
