+++ rtmlib/src/camera.cpp
@@ -10,13 +10,13 @@ using namespace rtm;
 Camera::Camera()
 {}
 
-Camera::Camera(const int& hsize, const int& vsize, const float& field_of_view)
+Camera::Camera(const int& hsize, const int& vsize, const double& field_of_view)
 : m_hsize(hsize), m_vsize(vsize), m_field_of_view(field_of_view), m_transform{Matrix(4, 4)}
 {
     m_transform.set_identity(); 
 
-    float half_view = tan(m_field_of_view/2);
-    float aspect = (static_cast<float>(m_hsize)/static_cast<float>(m_vsize));
+    double half_view = tan(m_field_of_view/2);
+    double aspect = (static_cast<double>(m_hsize)/static_cast<double>(m_vsize));
 
     if(aspect >= 1)
     {
@@ -35,12 +35,12 @@ Camera::Camera(const int& hsize, const int& vsize, const float& field_of_view)
 Ray Camera::ray_for_pixel(const int& px, const int& py) const // px and py are in canvas space (pixel locations)
 {
     // the offset from the edge of the canvas to the pixel's center
-    float xoffset = (px + 0.5f) * pixel_size;
-    float yoffset = (py + 0.5f) * pixel_size;
+    double xoffset = (px + 0.5f) * pixel_size;
+    double yoffset = (py + 0.5f) * pixel_size;
 
     // the untransformed coordinates of the pixel in world space.
-    float world_x = half_width - xoffset;
-    float world_y = half_height - yoffset;
+    double world_x = half_width - xoffset;
+    double world_y = half_height - yoffset;
 
     // using the camera matrix, transform the canvas point and the origin,
     // and then compute the ray's direction vector. (canvas is at z=-1)
