+++ rtmlib/src/intersection.cpp
@@ -5,12 +5,12 @@ using namespace rtm;
 Intersection::Intersection() : m_object(nullptr)
 {}
 
-Intersection::Intersection(const float& t_val , const Surface* object_val) : m_t(t_val)
+Intersection::Intersection(const double& t_val , const Surface* object_val) : m_t(t_val)
 {
     m_object = const_cast<Surface*>(object_val);
 }
 
-float Intersection::get_t() const
+double Intersection::get_t() const
 {
     return m_t;
 }
