+++ rtmlib/src/sphere.cpp
@@ -24,11 +24,11 @@ std::vector<Intersection> Sphere::intersects_with(const Ray& r) const
     
     Vector sphere_to_ray = ray2.origin() - Point(0, 0, 0);
 
-    float a = dot(ray2.direction(), ray2.direction());
-    float b = 2 * dot(ray2.direction(), sphere_to_ray);
-    float c = dot(sphere_to_ray, sphere_to_ray) - 1;
+    double a = dot(ray2.direction(), ray2.direction());
+    double b = 2 * dot(ray2.direction(), sphere_to_ray);
+    double c = dot(sphere_to_ray, sphere_to_ray) - 1;
 
-    float discriminant = pow(b, 2) - 4 * a * c;
+    double discriminant = std::pow(b, 2) - 4 * a * c;
 
     if(discriminant < 0)
     {
@@ -36,8 +36,8 @@ std::vector<Intersection> Sphere::intersects_with(const Ray& r) const
         return e;
     }
 
-    float t1 = (-b - sqrt(discriminant)) / (2 * a);
-    float t2 = (-b + sqrt(discriminant)) / (2 * a);
+    double t1 = (-b - std::sqrt(discriminant)) / (2 * a);
+    double t2 = (-b + std::sqrt(discriminant)) / (2 * a);
 
     std::vector<Intersection> intersections = {Intersection(t1, this), Intersection(t2, this)};
